requirements.txt
yfinance
numpy
pandas
matplotlib
google-generativeai